Add files via upload
In this problem are tasked with predicting the fare amount for a taxi ride in New York City given the pickup and dropoff cpLocations. While you can get a basic estimate based on just the distance between the two points, this will result in an RMSE of 5 to 8 dollars, depending on the model used.
This dataset with about 55 million records. *The data contains only the pickup date & time, the latitude & longitude (GPS coordinates) of the pickup and dropoff locations, and the number of passengers.*

- It is up to you to extract any further information. For instance, **does the time of day matter? The day of the week? How do we determine the distance traveled from pairs of GPS coordinates?**

For this data we've whittled the dataset down to just 120,000 records from April 11 to April 24, 2010.